+++ example/lib/widgets/auto_refresh.dart
@@ -8,21 +8,20 @@ class AutoRefresh extends StatefulWidget {
   final Widget child;
 
   AutoRefresh({
-    Key key,
-    @required this.duration,
-    @required this.child,
-  })  : assert(duration != null),
-        super(key: key);
+    Key? key,
+    required this.duration,
+    required this.child,
+  }) : super(key: key);
 
   @override
   _AutoRefreshState createState() => _AutoRefreshState();
 }
 
 class _AutoRefreshState extends State<AutoRefresh> {
-  int keyValue;
-  ValueKey key;
+  int? keyValue;
+  ValueKey? key;
 
-  Timer _timer;
+  Timer? _timer;
 
   @override
   void initState() {
@@ -47,7 +46,7 @@ class _AutoRefreshState extends State<AutoRefresh> {
       widget.duration,
       () {
         setState(() {
-          keyValue = keyValue + 1;
+          keyValue = keyValue! + 1;
           key = ValueKey(keyValue);
           _recursiveBuild();
         });